bench/ibench/requirements.txt
torch==1.10.0
numpy
pandas
psycopg2-binary
